requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
langsmith
python-dotenv
unstructured[pdf,pptx,docx]
o365
psycopg-binary
openai
psycopg[binary,pool]
sse-starlette
pydantic
pytest
numpy
azure-cosmos
langgraph
openpyxl